\name{Fit Methods}
\docType{methods}
\alias{fit-methods}
\alias{fit,StatModel,LearningSample-method}
\title{ Fit `StatModel' Objects to Data }
\description{

    Fit a `StatModel' model to objects of class `LearningSample'.

}
\section{Methods}{
  \describe{
    \item{fit}{\code{signature(model = "StatModel", data = "LearningSample")}:
               fit \code{model} to \code{data}.}
  }
}
\keyword{methods}
