\name{Initialize Methods}
\docType{methods}
\alias{initialize}
\alias{initialize-methods}
\alias{initialize,ExpectCovarInfluence-method}
\alias{initialize,ExpectCovar-method}
\alias{initialize,LinStatExpectCovar-method}
\alias{initialize,LinStatExpectCovarMPinv-method}
\alias{initialize,svd_mem-method}
\alias{initialize,VariableFrame-method}
\title{ Methods for Function initialize in Package `party' }
\description{
  Methods for function \code{initialize} in package \pkg{party} -- those are
  internal functions not to be called by users.
}
\section{Methods}{
\describe{

\item{.Object = "ExpectCovarInfluence"}{\code{new("ExpectCovarInfluence")}}

\item{.Object = "ExpectCovar"}{\code{new("ExpectCovar")}}

\item{.Object = "LinStatExpectCovar"}{\code{new("LinStatExpectCovar")}}

\item{.Object = "LinStatExpectCovarMPinv"}{\code{new("LinStatExpectCovarMPinv")}}

\item{.Object = "VariableFrame"}{\code{new("VariableFrame")}}

}}
\keyword{methods}
