\name{initVariableFrame-methods}
\docType{methods}
\alias{initVariableFrame}
\alias{initVariableFrame-methods}
\alias{initVariableFrame,data.frame-method}
\alias{initVariableFrame,matrix-method}
\title{Set-up VariableFrame objects}
\description{
 Set-up VariableFrame objects
}
\section{Methods}{
These methods are not to be called by the user.
\describe{

\item{\code{signature(obj = "data.frame")}}{
converges a data frame to VariableFrame
}

\item{\code{signature(obj = "matrix")}}{
converges a matrix to VariableFrame
}
}}
\keyword{methods}
