\name{SplittingNode Class}
\docType{class}
\alias{SplittingNode-class}
\alias{TerminalNode-class}
\alias{TerminalModelNode-class}

\title{Class "SplittingNode"}
\description{  
    A list representing the inner node of a binary tree.
}
\section{Extends}{
Class \code{"list"}, from data part.
Class \code{"vector"}, by class \code{"list"}. See
\code{\link{BinaryTree-class}} for more details.
}
\keyword{classes}
