\name{TreeControl Class}
\docType{class}
\alias{TreeControl-class}
\alias{TreeControl}

\title{Class "TreeControl"}
\description{ 
   Objects of this class represent the hyper parameter setting for tree
   growing.
}
\section{Objects from the Class}{
Objects can be created by \code{\link{ctree_control}}.
}
\section{Slots}{
  \describe{
    \item{\code{varctrl}:}{Object of class \code{"VariableControl"}.}
    \item{\code{splitctrl}:}{Object of class \code{"SplitControl"}.}
    \item{\code{gtctrl}:}{Object of class \code{"GlobalTestControl"}.}
    \item{\code{tgctrl}:}{Object of class \code{"TreeGrowControl"}.}
  }
}
\section{Methods}{
No methods defined with class "TreeControl" in the signature.
}
\keyword{classes}
